refactor typescript ok
