+++ opencti-platform/opencti-graphql/package.json
@@ -172,7 +172,6 @@
     "@types/bluebird": "3.5.42",
     "@types/dot-object": "2.1.6",
     "@types/ejs": "3.1.5",
-    "@types/eventsource": "1.1.15",
     "@types/express": "4.17.21",
     "@types/graphql-upload": "16.0.7",
     "@types/mime-types": "2.1.4",
